+++ packages/orq-rc/RELEASES.md
@@ -10939,3 +10939,13 @@ Based on:
 - [python v4.15.0-rc.47] packages/orq-rc
 ### Releases
 - [PyPI v4.15.0-rc.47] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.47 - packages/orq-rc
+
+## 2026-09-15 09:36:50
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.797.0 (2.937.18) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.15.0-rc.48] packages/orq-rc
+### Releases
+- [PyPI v4.15.0-rc.48] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.48 - packages/orq-rc